contrib/castoretpollux/stablediffusion/requirements.txt
accelerate==0.27.2
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.26.3
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
invisible-watermark==0.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psutil==5.9.8
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyWavelets==1.5.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
zipp==3.17.0
